AddReminder a new way for adding reminders moved from route to component left the action for submitting ---- SafePoint ---- handling date to display localDatetime
